src/backend/requirements.txt
pandas
fastapi
uvicorn
asyncio
pyserial
python-socketio
